Added CMake presets, errors to stderr

+++ msdf-atlas-gen/image-save.hpp
@@ -69,17 +69,17 @@ template <int N>
 bool saveImageBinary(const msdfgen::BitmapConstRef<byte, N> &bitmap, const char *filename, YDirection outputYDirection) {
     bool success = false;
     if (FILE *f = fopen(filename, "wb")) {
-        int written = 0;
+        size_t written = 0;
         switch (outputYDirection) {
             case YDirection::BOTTOM_UP:
-                written = fwrite(bitmap.pixels, 1, N*bitmap.width*bitmap.height, f);
+                written = fwrite(bitmap.pixels, 1, (size_t) N*bitmap.width*bitmap.height, f);
                 break;
             case YDirection::TOP_DOWN:
                 for (int y = bitmap.height-1; y >= 0; --y)
-                    written += fwrite(bitmap.pixels+N*bitmap.width*y, 1, N*bitmap.width, f);
+                    written += fwrite(bitmap.pixels+(size_t) N*bitmap.width*y, 1, (size_t) N*bitmap.width, f);
                 break;
         }
-        success = written == N*bitmap.width*bitmap.height;
+        success = written == (size_t) N*bitmap.width*bitmap.height;
         fclose(f);
     }
     return success;
@@ -95,17 +95,17 @@ bool
         (const msdfgen::BitmapConstRef<float, N> &bitmap, const char *filename, YDirection outputYDirection) {
     bool success = false;
     if (FILE *f = fopen(filename, "wb")) {
-        int written = 0;
+        size_t written = 0;
         switch (outputYDirection) {
             case YDirection::BOTTOM_UP:
-                written = fwrite(bitmap.pixels, sizeof(float), N*bitmap.width*bitmap.height, f);
+                written = fwrite(bitmap.pixels, sizeof(float), (size_t) N*bitmap.width*bitmap.height, f);
                 break;
             case YDirection::TOP_DOWN:
                 for (int y = bitmap.height-1; y >= 0; --y)
-                    written += fwrite(bitmap.pixels+N*bitmap.width*y, sizeof(float), N*bitmap.width, f);
+                    written += fwrite(bitmap.pixels+(size_t) N*bitmap.width*y, sizeof(float), (size_t) N*bitmap.width, f);
                 break;
         }
-        success = written == N*bitmap.width*bitmap.height;
+        success = written == (size_t) N*bitmap.width*bitmap.height;
         fclose(f);
     }
     return success;
@@ -121,9 +121,9 @@ bool
         (const msdfgen::BitmapConstRef<float, N> &bitmap, const char *filename, YDirection outputYDirection) {
     bool success = false;
     if (FILE *f = fopen(filename, "wb")) {
-        int written = 0;
+        size_t written = 0;
         for (int y = 0; y < bitmap.height; ++y) {
-            const float *p = bitmap.pixels+N*bitmap.width*(outputYDirection == YDirection::TOP_DOWN ? bitmap.height-y-1 : y);
+            const float *p = bitmap.pixels+(size_t) N*bitmap.width*(outputYDirection == YDirection::TOP_DOWN ? bitmap.height-y-1 : y);
             for (int x = 0; x < bitmap.width; ++x) {
                 const unsigned char *b = reinterpret_cast<const unsigned char *>(p++);
                 for (int i = sizeof(float)-1; i >= 0; --i)
@@ -143,7 +143,7 @@ bool saveImageText(const msdfgen::BitmapConstRef<byte, N> &bitmap, const char *f
     if (FILE *f = fopen(filename, "wb")) {
         success = true;
         for (int y = 0; y < bitmap.height; ++y) {
-            const byte *p = bitmap.pixels+N*bitmap.width*(outputYDirection == YDirection::TOP_DOWN ? bitmap.height-y-1 : y);
+            const byte *p = bitmap.pixels+(size_t) N*bitmap.width*(outputYDirection == YDirection::TOP_DOWN ? bitmap.height-y-1 : y);
             for (int x = 0; x < N*bitmap.width; ++x)
                 success &= fprintf(f, x ? " %02X" : "%02X", (unsigned) *p++) > 0;
             success &= fprintf(f, "\n") > 0;
@@ -159,7 +159,7 @@ bool saveImageText(const msdfgen::BitmapConstRef<float, N> &bitmap, const char *
     if (FILE *f = fopen(filename, "wb")) {
         success = true;
         for (int y = 0; y < bitmap.height; ++y) {
-            const float *p = bitmap.pixels+N*bitmap.width*(outputYDirection == YDirection::TOP_DOWN ? bitmap.height-y-1 : y);
+            const float *p = bitmap.pixels+(size_t) N*bitmap.width*(outputYDirection == YDirection::TOP_DOWN ? bitmap.height-y-1 : y);
             for (int x = 0; x < N*bitmap.width; ++x)
                 success &= fprintf(f, x ? " %g" : "%g", *p++) > 0;
             success &= fprintf(f, "\n") > 0;
